Add the TracingMacros module and the `@Traced` macro
**Motivation:**

Adding tracing to functions can be a slightly invasive modification
because using `withSpan` requires re-indenting the whole function. It
also requires some duplicated modifications when modifying the effects
signatures of functions. We want a macro to perform this transformation.

**Modifications:**

Introduce a TracingMacros module to provide macros for the Tracing
package. Add a `@Traced` macro that performs the transformation.

This is defined as a separate product so that users who don't want to
pay the compile-time cost of macros don't have to use it, you opt-in to
that cost by depending on the TracingMacros module.

The `@Traced` macro is only available in Swift 6.0 compilers since
function body macros were introduced in Swift 6.0.