Support async/throws functions in the `@Traced` macro
**Motivation:**

We need to support `@Traced` on async and throws functions to make it
compatible with everywhere people want to use them.

**Modifications:**

Based on the effects signature of the attached function we apply
try/await as appropriate. But, if the function is async/throws but those
effects aren't actually used in the function body, this causes a new
warning because the closure isn't inferred to be async and/or throws. To
avoid those warnings, we also apply matching effects specifiers to the
withSpan closure.

**Result:**

The `@Traced` macro is now usable on all functions.