Add support for parameters passed to the `@Traced` macro
**Motivation:**

The `@Traced` macro isn't sufficient for all cases of `withSpan` because
it doesn't allow the same customization of the span. We need to expose
those parameters to the macro.

**Modifications:**

First, allow overriding the regular parameters of the withSpan call:
setting the operationName, the context, and the kind. If it's not
specified, it's not passed to the function, which means the function
remains the source of truth for default arguments.

Also allow overriding the span name, which controls the variable binding
in the closure body. This is primarily useful for avoiding shadowing an
outer "span" variable.

**Result:**

Users of the macro can use the operationName, context, and kind
parameters to customize the span like `withSpan`, and can use the span
parameter to customize the variable introduced by the macro's expansion.